+++ dop-web/src/pages/Code/index.js
@@ -1,4 +1,4 @@
 import Code from './DataOverview'
-import {PersonalProjects,StarredProjects,AllProjects} from './projects'
+import {PersonalProjects,StarredProjects,AllProjects} from './Projects'
 import ProjectOverview from './ProjectOverview'
 export {Code,PersonalProjects,StarredProjects,AllProjects,ProjectOverview};
